unittests/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
-e file:../core_lib
    # via -r requirements.in
aiohttp==3.7.3
    # via -r requirements.in
appdirs==1.4.4
    # via black
astroid==2.4.2
    # via pylint
async-timeout==3.0.1
    # via aiohttp
asynctest==0.13.0
    # via -r requirements.in
attrs==20.3.0
    # via
    #   aiohttp
    #   flake8-bugbear
    #   pytest
bandit==1.7.0
    # via flake8-bandit
black==20.8b1
    # via -r requirements.in
chardet==3.0.4
    # via aiohttp
click==7.1.2
    # via black
coverage==5.3.1
    # via
    #   -r requirements.in
    #   pytest-cov
faker==5.4.0
    # via -r requirements.in
flake8-assertive==1.3.0
    # via -r requirements.in
flake8-bandit==2.1.2
    # via -r requirements.in
flake8-blind-except==0.2.0
    # via -r requirements.in
flake8-bugbear==20.11.1
    # via -r requirements.in
flake8-builtins==1.5.3
    # via -r requirements.in
flake8-comprehensions==3.3.1
    # via -r requirements.in
flake8-functions==0.0.4
    # via -r requirements.in
flake8-logging-format==0.6.0
    # via -r requirements.in
flake8-polyfill==1.0.2
    # via flake8-bandit
flake8-print==4.0.0
    # via -r requirements.in
flake8-string-format==0.3.0
    # via -r requirements.in
flake8-tidy-imports==4.2.1
    # via -r requirements.in
flake8-todo==0.7
    # via -r requirements.in
flake8-variables-names==0.0.3
    # via -r requirements.in
flake8==3.8.4
    # via
    #   -r requirements.in
    #   flake8-assertive
    #   flake8-bandit
    #   flake8-bugbear
    #   flake8-builtins
    #   flake8-comprehensions
    #   flake8-polyfill
    #   flake8-print
    #   flake8-string-format
    #   flake8-tidy-imports
gitdb==4.0.5
    # via gitpython
gitpython==3.1.12
    # via bandit
idna==3.1
    # via yarl
iniconfig==1.1.1
    # via pytest
isort==4.3.4
    # via
    #   -r requirements.in
    #   pylint
lazy-object-proxy==1.4.3
    # via astroid
mccabe==0.6.1
    # via
    #   flake8
    #   pylint
multidict==5.1.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
mypy==0.790
    # via -r requirements.in
packaging==20.8
    # via pytest
pathspec==0.8.1
    # via black
pbr==5.5.1
    # via stevedore
pluggy==0.13.1
    # via pytest
py==1.10.0
    # via pytest
pycodestyle==2.6.0
    # via
    #   flake8
    #   flake8-bandit
    #   flake8-print
    #   flake8-todo
pyflakes==2.2.0
    # via flake8
pyhamcrest==2.0.2
    # via -r requirements.in
pylint==2.6.0
    # via -r requirements.in
pyparsing==2.4.7
    # via packaging
pytest-asyncio==0.14.0
    # via -r requirements.in
pytest-cov==2.10.1
    # via -r requirements.in
pytest==6.2.1
    # via
    #   -r requirements.in
    #   flake8-functions
    #   pytest-asyncio
    #   pytest-cov
python-dateutil==2.8.1
    # via faker
pytz==2020.5
    # via -r requirements.in
pyyaml==5.3.1
    # via bandit
regex==2020.11.13
    # via black
six==1.15.0
    # via
    #   astroid
    #   bandit
    #   flake8-print
    #   python-dateutil
smmap==3.0.4
    # via gitdb
stevedore==3.3.0
    # via bandit
text-unidecode==1.3
    # via faker
toml==0.10.2
    # via
    #   black
    #   pylint
    #   pytest
typed-ast==1.4.2
    # via
    #   black
    #   mypy
typing-extensions==3.7.4.3
    # via
    #   aiohttp
    #   black
    #   mypy
wrapt==1.12.1
    # via astroid
yarl==1.6.3
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
